+++ cloudsim-plus-testbeds/src/main/java/org/cloudsimplus/sla/readJsonFile/SlaMetricDimension.java
@@ -31,7 +31,11 @@ package org.cloudsimplus.sla.readJsonFile;
  * 
  * Each dimension contains the name of the metric, the minimum and maximum 
  * acceptable value, and the metric unit. 
+<<<<<<< HEAD:cloudsim-plus-testbeds/src/main/java/org/cloudsimplus/sla/readJsonFile/SlaMetricDimension.java
  * See: <http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/viewing_metrics_with_cloudwatch.html>
+=======
+ * See: http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/viewing_metrics_with_cloudwatch.html
+>>>>>>> upstream/master:cloudsim-plus-testbeds/src/main/java/org/cloudsimplus/sla/readJsonFile/SlaMetricDimension.java
  * 
  * @author raysaoliveira
  */