+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/distributions/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Provides Pseudo Random Number Generators (PRNG) following several statistical
+ * distributions used by the simulation API.
+ * Additionally, they can be used by developers implementing their own simulations.
+ *
+ * <p>The most basic PRNG is the {@link org.cloudbus.cloudsim.distributions.UniformDistr}
+ * that generates uniformly distributed pseudo random numbers.</p>
+ */
+package org.cloudbus.cloudsim.distributions;