+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/lists/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Provides a set of helper class to manage lists of different objects
+ * such as Hosts, VMs, Cloudlets and PEs. Such helper classes are used
+ * internally by CloudSim Plus.
+ */
+package org.cloudbus.cloudsim.lists;