+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/schedulers/cloudlet/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * The package provides {@link org.cloudbus.cloudsim.schedulers.cloudlet.CloudletScheduler}
+ * implementations that are used to schedule the execution of multiple
+ * {@link org.cloudbus.cloudsim.cloudlets.Cloudlet} inside a given
+ * {@link org.cloudbus.cloudsim.vms.Vm}.
+ *
+ * <p>See the documentation of the package from the upper level for more details.</p>
+ *
+ * @author Manoel Campos da Silva Filho
+ */
+package org.cloudbus.cloudsim.schedulers.cloudlet;