+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/schedulers/vm/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * The package provides {@link org.cloudbus.cloudsim.schedulers.vm.VmScheduler}
+ * implementations that are used to schedule the execution of multiple
+ * {@link org.cloudbus.cloudsim.vms.Vm} inside a given
+ * {@link org.cloudbus.cloudsim.hosts.Host}.
+ *
+ * <p>See the documentation of the package from the upper level for more details.</p>
+ *
+ * @author Manoel Campos da Silva Filho
+ */
+package org.cloudbus.cloudsim.schedulers.vm;