+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/utilizationmodels/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * Provides classes that model utilization of resources such as
+ * {@link org.cloudbus.cloudsim.resources.Pe CPU}, {@link org.cloudbus.cloudsim.resources.Ram} and
+ * {@link org.cloudbus.cloudsim.resources.Bandwidth},
+ * defining how a given resource is used  by a {@link org.cloudbus.cloudsim.cloudlets.Cloudlet}
+ * along the simulation time.
+ *
+ * <p>The most basic utilization model that can be used for any of the mentioned resources
+ * is the {@link org.cloudbus.cloudsim.utilizationmodels.UtilizationModelFull}.</p>
+ *
+ * @author Manoel Campos da Silva Filho
+ */
+package org.cloudbus.cloudsim.utilizationmodels;